Flowchart Tooltip Trigger
